+++ NexUs/Source/DeveloperComponents/NXCheckBoxStyle.cpp
@@ -41,7 +41,7 @@ NXCheckBoxStyle::drawControl(ControlElement element,
       int yOffset        = (checkBoxRect.height() - _pCheckIndicatorWidth) / 2;
       QRect checkRect(checkBoxRect.x(), checkBoxRect.y() + yOffset, _pCheckIndicatorWidth, _pCheckIndicatorWidth);
       checkRect.adjust(1, 1, -1, -1);
-      // 复选框绘制
+      //复选框绘制
       painter->setPen(Qt::NoPen);
       if (bopt->state.testFlag(QStyle::State_On) || bopt->state.testFlag(QStyle::State_NoChange))
       {
@@ -83,7 +83,7 @@ NXCheckBoxStyle::drawControl(ControlElement element,
         }
       }
       painter->drawRoundedRect(checkRect, _pBorderRadius, _pBorderRadius);
-      // 图标绘制
+      //图标绘制
       painter->setPen(NXThemeColor(NXThemeType::Dark, BasicText));
       if (bopt->state.testFlag(QStyle::State_On))
       {
@@ -91,7 +91,7 @@ NXCheckBoxStyle::drawControl(ControlElement element,
         QFont iconFont = QFont(QStringLiteral("NXAwesome"));
         iconFont.setPixelSize(_pCheckIndicatorWidth * 0.75);
         painter->setFont(iconFont);
-        painter->drawText(checkRect, Qt::AlignCenter, QChar(NXIconType::Check));
+        painter->drawText(checkRect, Qt::AlignCenter, QChar((unsigned short) NXIconType::Check));
         painter->restore();
       }
       else if (bopt->state.testFlag(QStyle::State_NoChange))
@@ -99,7 +99,7 @@ NXCheckBoxStyle::drawControl(ControlElement element,
         QLine checkLine(checkRect.x() + 3, checkRect.center().y(), checkRect.right() - 3, checkRect.center().y());
         painter->drawLine(checkLine);
       }
-      // 文字绘制
+      //文字绘制
       painter->setPen(isEnabled ? NXThemeColor(_themeMode, BasicText) : NXThemeColor(_themeMode, BasicTextDisable));
       QRect textRect(checkRect.right() + 10, checkBoxRect.y() + yOffset, checkBoxRect.width() - checkRect.right() - 10,
                      _pCheckIndicatorWidth);
